Extract case class serialization in separate trait
